\name{funpca-internal}
\alias{dbind}
\alias{funpca.default}
\alias{funpcaEst}
\alias{spectralnorm}
\alias{plot.funpcaa}
\alias{print.funpca}
\alias{summary.funpcaa}
\alias{Sqrt}

\title{Internal fpcamm objects}
\description{Internal functions}
\details{These are not to be called by the user}
\keyword{internal}
